% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillotest-package.R
\docType{package}
\name{armadillotest-package}
\alias{armadillotest}
\alias{armadillotest-package}
\title{armadillotest: Testing armadillo}
\description{
Testing armadillo. I run this within a tiny package to test the compiled code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/armadillo}
  \item Report bugs at \url{https://github.com/pachadotdev/armadillo/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
